Reorganized SimpleGLView,Renderer and Fragment.  This fixed the issue of the app crashing when moving into the background and then back to foreground.  This also move standard gl management code out of the renderers and into a common place.
ManGL no longer contains an assetmanager.  This was a liability as it potentially could keep anything related to the Context in memory.

Changed main activity to not destroy itself on screen size/orientation change.  For the examples this doesn't matter, but did it to test making the SimpleGLView more efficient.